refactor: ProfileInfoCard와 ProfileInfoCard 도메인 변경
